requirements.txt
langchain>=0.3.0
langchain-mistralai>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.0
yfinance>=0.2.40
python-dotenv>=1.0.0
pydantic>=2.0.0
